+++ src/ch/unizh/ini/caviar/chip/retina/Tmpdiff128.java
@@ -96,7 +96,7 @@ public class Tmpdiff128 extends AERetina implements Serializable {
                 }
             }
             int sxm=sizeX-1;
-            short[] a=in.getAddresses();
+            int[] a=in.getAddresses();
             int[] timestamps=in.getTimestamps();
             OutputEventIterator outItr=out.outputIterator();
             for(int i=0;i<n;i+=skipBy){ // bug here
